reward amount u32

+++ dojo/src/models/summit.cairo
@@ -42,7 +42,7 @@ pub struct RewardEvent {
     #[key]
     pub beast_token_id: u32,
     pub owner: ContractAddress,
-    pub amount: u128,
+    pub amount: u32,
 }
 
 #[derive(Drop, Serde)]
